Rename CoinTextBehaviour and CoinText2Behaviour to CoinUIBehaviour and CoinUI2Behaviour respectively
This commit updates the class names `CoinTextBehaviour` and `CoinText2Behaviour` to `CoinUIBehaviour` and `CoinUI2Behaviour` in the respective C# scripts. The change reflects a shift in the responsibilities of these classes to handle user interface (UI) tasks rather than just text updating. Consequently, the respective scripts' filenames have been updated accordingly.

In addition, the changes impact the `CoinBehaviour` and `Coin2Behaviour` classes that previously relied on `CoinTextBehaviour` and `CoinText2Behaviour`. Now, they interact with `CoinUIBehaviour` and `CoinUI2Behaviour` to update the